--- router/usecase/pools/constants.go
@@ -1,12 +1,0 @@
-package pools
-
-import (
-	"math/big"
-
-	"github.com/osmosis-labs/osmosis/osmomath"
-)
-
-var (
-	oneDec    = osmomath.OneDec()
-	oneDecBig = big.NewFloat(1)
-)
